backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
drf-spectacular==0.26.5
psycopg2-binary>=2.9.10
python-decouple==3.8
Pillow>=11.3.0
celery>=5.5.3
redis>=6.2.0
pytest>=8.4.1
pytest-django>=4.11.1
pytest-cov>=6.2.1
black>=25.1.0
flake8>=7.3.0
isort>=6.0.1
pre-commit>=4.2.0
gunicorn>=23.0.0
whitenoise>=6.9.0 